srgan/requirements.txt
torch
torchvision
numpy
pillow
tqdm
matplotlib
albumentations